버튼 옵션을 for문으로 바꿈. getColor와 getText가 별도로 존재
